fix for Browser class
